requirements.txt
python-dotenv==0.20.0
python-telegram-bot==20.0
pyrogram
tgcrypto==1.2.3
dnspython==2.3.0
motor==3.1.1
fastapi
uvicorn
